+++ example/simlite.py
@@ -1,0 +1,45 @@
+from pyhcl import *
+from pysv import sv, DataType, Reference
+from pyhcl.simulator import Simlite, DpiConfig
+
+class Add(BlackBox):
+    io = IO(
+        in1=Input(U.w(32)),
+        in2=Input(U.w(32)),
+        out=Output(U.w(32))
+    )
+
+
+@sv(a=DataType.UInt, b=DataType.UInt, return_type=Reference(x = DataType.UInt))
+def fn(a, b):
+    return a + b
+
+addpysvmodule(Add, fn)
+compile_and_binding_all()
+
+
+
+class Top(Module):
+    io = IO(
+        a=Input(U.w(32)),
+        b=Input(U.w(32)),
+        c=Output(U.w(32))
+    )
+
+    add = Add()
+    add.io.in1 <<= io.a
+    add.io.in2 <<= io.b
+    io.c <<= add.io.out
+
+
+from random import randint
+
+if __name__ == '__main__':
+    cfg = DpiConfig()
+    #Emitter.dumpVerilog(Emitter.dump(Emitter.emit(Top()), "Top.fir"))
+
+    s = Simlite(Top(), cfg)
+    s.step([20, 20])
+    s.step([15, 10])
+    s.step([1000, 1])
+    s.step([999, 201])
